Create a new helper file for methods realted to focus
